+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added context-sensitive crosshair
 - Announce alias changes to all clients.
+- Dance animation
 
 ### Changed
 
 - Improved camera aiming
 - Made civsim, sites, etc. deterministic from the same seed.
+- Improved animations by adding orientation variation
 
 ### Removed